[Fix] Adds a `delete` operation and a required `@id` query parameter to collection-valued nav. property paths with `$ref` (#477)
* Add delete operation to col-valued nav. paths with $ref

We no longer use containment to determine whether or not to create $ref segments for nav. props.; we use referenceable annotations. Hence the Debug.Assert needs omission.
The query parameter @id needs to be required in the delete operation of col-valued nav. props. with $ref segments because you need to index into the collection before deleting

* Update description

* Update tests and integration files

* Update release notes

* Update release note description